Refactor(AAS models): split AAS models and add Catena-X specific models file (#65)
* refactor: move AAS models to aas folder and split aas models to base and dto

* feat: base file for catenax specific aas models

* refactor: rename get_version_info to get_version and update return type; add VersionedModel classes that didn't inherit
