+++ src/tractusx_sdk/industry/models/aas/v3/catenax.py
@@ -11,7 +11,7 @@
 # https://www.apache.org/licenses/LICENSE-2.0.
 #
 # Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS
+# distributed under the License is distributed on an "AS IS" BASIS,
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
 # either express or implied. See the
 # License for the specific language govern in permissions and limitations
@@ -20,6 +20,5 @@
 # SPDX-License-Identifier: Apache-2.0
 #################################################################################
 
-## Where the models are stored
-
-from .example import ParentExample, ChildExample
+# This file provides models specific for Catena-X Use Cases
+# Here you can define templates that are specific to the Catena-X use cases
